requirements.txt
# Task Manager Desktop App — Dependencies
# Python 3.10+

# UI Framework
flet>=0.21.0

# Database ORM
SQLAlchemy>=2.0.0

# Database Migration
alembic>=1.13.0

# Date utilities
python-dateutil>=2.9.0

# Excel export
openpyxl>=3.1.0

# PDF export
reportlab>=4.0.0

# Word export (diary)
python-docx>=1.1.0

# Charts (dashboard)
matplotlib>=3.7.0

# FastAPI server (Phase 21)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# Testing
pytest>=9.0.0
pytest-cov>=7.0.0
